move functions to useOnramp() hook
